requirements.txt
Flask
torch
torchvision
numpy
opencv-python
PyYAML
safetensors
scipy
scikit-image
matplotlib